crypto_prediction/requirements.txt
tensorflow>=2.8.0
keras>=2.8.0
pandas>=1.4.0
numpy>=1.22.0
scikit-learn>=1.0.0
matplotlib>=3.5.0
seaborn>=0.11.0
prophet>=1.0.0
xgboost>=1.5.0
statsmodels>=0.13.0
requests>=2.27.0
python-binance>=1.0.0
fastapi>=0.75.0
uvicorn>=0.17.0
plotly>=5.6.0
dash>=2.3.0
pytest>=7.0.0
python-dotenv>=0.20.0
pyyaml>=6.0.0
joblib>=1.1.0
ta>=0.10.0  # Technical analysis library
tqdm>=4.64.0
